I added in seperate folders for my extensions and also changed the name of the one folder from tbale logic to table cell
