+++ Paws4Me/Paws4Me/Controllers/Logic/ViewController.swift
@@ -37,7 +37,6 @@ class ViewController: UIViewController, UISearchBarDelegate {
            self.title = "Adoptable Animals"
            setUpSearchbar()
        }
-    
     func getAdoption() {
         let urlString = Constants.adoptURL
         if let request = URL(string: urlString) {
